smth

+++ backend/routes/blogs.js
@@ -47,13 +47,27 @@ router.post('/', async (request, response) => {
     })
 
     const savedBlog = await blog.save()
+    const populatedBlog = await savedBlog.populate('user', { username: 1, name: 1})
+
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
 
-    response.status(201).json(savedBlog)
+    response.status(201).json(populatedBlog)
 })
 
 router.delete('/:id', async (request, response) => {
+    const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        return response.status(404).json({ error: 'blog not found' })
+    }
+    if (blog.user.toString() !== decodedToken.id) {
+        return response.status(403).json({ error: 'unauthorized deletion' })
+    }
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
